services/data/Docker.requirements.txt
weaviate-client
openai==0.28.1
faiss-cpu
langchain==0.0.327
tiktoken
selenium
unstructured
newspaper3k
PyPDF2
beautifulsoup4
lxml==4.9.2
pycryptodome
boto3
tenacity
pyyaml
pandas
requests
unittest-parametrize
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-docx